requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.2
sentence-transformers==2.5.1
plotly==5.18.0
scikit-learn==1.3.2
pytest==7.4.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyspark==3.5.0
httpx==0.25.2
huggingface_hub>=0.20.0
python-dotenv==1.0.0
loguru==0.7.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2